Move the functions of the internal namespace of vector.templates to their own file.
